Check for `definitions` block (+ refactoring)
