+++ src/completion.js
@@ -1,48 +1,23 @@
 "use strict"
 
 const vscode = require('vscode')
+const common = require('./common')
 const { attributes, blocks, functions, variables } = require('./schema')
 
 const selector = { language: 'couper' }
 
-const parentBlockRegex = /\b([\w-]+)(?:[ \t]+"[^"]+")?[ \t]*=?[ \t]*{[^{}]*$/
-const blockRegex = /{[^{}]*}/g
 const attributeRegex = /^\s*"?\(?([\w-]+)\)?"?\s*=/
 const attributeMustRegex = /^\s*"?\(?([\w-]+)\)?"?\s*=$/
-// see http://regex.info/listing.cgi?ed=2&p=281
-const filterRegex = /([^"/#]+|"(?:\\.|[^"\\])*")|\/\*[^*]*\*+(?:[^/*][^*]*\*+)*\/|(?:\/\/|#)[^\n]*/g
 const variableRegex = /(.+)\..*\.$/
 
 const providers = []
-
-function getParentBlock(document, position) {
-	const range = new vscode.Range(document.positionAt(0), position)
-	let text = document.getText(range)
-
-	text = text.replace(filterRegex, (match, match1) => {
-		if (match1 === undefined) {
-			return "" // Comment
-		}
-		if (match1[0] === '"') {
-			return '"..."'
-		}
-		return match1
-	})
-
-	while (blockRegex.test(text)) {
-		text = text.replace(blockRegex, "")
-	}
-
-	const matches = text.match(parentBlockRegex)
-	return matches ? matches[1] : ""
-}
 
 for (const [name, block] of Object.entries(blocks)) {
 	const provider = vscode.languages.registerCompletionItemProvider(selector,
 		{
 			provideCompletionItems(document, position, token, context) {
 				const linePrefix = document.lineAt(position).text.substr(0, position.character)
-				const parentBlock = getParentBlock(document, position)
+				const parentBlock = common.getParentBlock(document, position)
 				if (!/^\s*([\w-]+)?$/.test(linePrefix) || (block.parents || ['']).indexOf(parentBlock) === -1) {
 					return undefined
 				}
@@ -85,7 +60,7 @@ for (const [name, attribute] of Object.entries(attributes)) {
 		{
 			provideCompletionItems(document, position, token, context) {
 				const linePrefix = document.lineAt(position).text.substr(0, position.character)
-				const parentBlock = getParentBlock(document, position)
+				const parentBlock = common.getParentBlock(document, position)
 				const writeAttrRegex = /^\s*([\w-]+)?$/
 				if (!writeAttrRegex.test(linePrefix) || (attribute.parents || []).indexOf(parentBlock) === -1) {
 					return undefined
