+++ src/extension.js
@@ -2,8 +2,8 @@
 
 Object.defineProperty(exports, "__esModule", { value: true })
 
-const Completion = require("./completion.js")
-const Definition = require("./definition.js")
+const Completion = require("./completion")
+const Definition = require("./definition")
 
 exports.activate = (context) => {
 	context.subscriptions.concat(Completion.providers, Definition.providers)
